requirements.txt
pandas
ipykernel
PyYAML
boto3
psycopg2
python-dotenv